Added HashMap for library, functional add and remove commands.

+++ src/model/PersonalLibrary.java
@@ -24,7 +24,14 @@ public class PersonalLibrary {
                 System.out.println("Library created");
             }
             else{
-                this.collection.put("", database.getPlayable(""));
+                BufferedReader reader;
+                reader = new BufferedReader(new FileReader(filename));
+                String line = reader.readLine();
+                while (line != null) {
+                    line = reader.readLine();
+                    this.collection.put(line, database.getPlayable(line));
+                }
+                reader.close();
             }
         } catch (IOException e) {
             e.printStackTrace();
@@ -32,37 +39,46 @@ public class PersonalLibrary {
 
     }
     private void writeToLibrary(Playable playable){
-        try {
-            FileWriter myWriter = new FileWriter("filename.txt");
-            myWriter.write(playable.getGUID() + "\n");
-            myWriter.close();
-        } catch (IOException e) {
-            e.printStackTrace();
+        if (!collection.containsKey(playable)) {
+            collection.put(playable.getGUID(), playable);
+            try {
+                FileWriter myWriter = new FileWriter("filename.txt");
+                myWriter.write(playable.getGUID() + "\n");
+                myWriter.close();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+        else{
+            System.out.println("Song/Release is already in your library.");
         }
     }
-    private void eraseFromLibrary(Playable playable) throws IOException {
-
-        File inputFile = new File(filename);
-        File tempFile = new File("TempFile.txt");
-
-        BufferedReader reader = new BufferedReader(new FileReader(inputFile));
-        BufferedWriter writer = new BufferedWriter(new FileWriter(tempFile));
-
+    private boolean eraseFromLibrary(Playable playable) throws IOException {
         String GUID = playable.getGUID();
-        String currentLine;
+        if (collection.remove(GUID, playable)) {
+            File inputFile = new File(filename);
+            File tempFile = new File("TempFile.txt");
 
-        while((currentLine = reader.readLine()) != null) {
-            // trim newline when comparing with lineToRemove
-            String trimmedLine = currentLine.trim();
-            if(trimmedLine.equals(GUID)) continue;
-            writer.write(currentLine + System.getProperty("line.separator"));
+            BufferedReader reader = new BufferedReader(new FileReader(inputFile));
+            BufferedWriter writer = new BufferedWriter(new FileWriter(tempFile));
+
+            String currentLine;
+
+            while ((currentLine = reader.readLine()) != null) {
+                String trimmedLine = currentLine.trim();
+                if (trimmedLine.equals(GUID)) continue;
+                writer.write(currentLine + System.getProperty("line.separator"));
+            }
+
+            writer.close();
+            reader.close();
+            inputFile.delete();
+            return tempFile.renameTo(inputFile);
         }
-
-        writer.close();
-        reader.close();
-        inputFile.delete();
-        boolean successful = tempFile.renameTo(inputFile);
-
+        else{
+            System.out.println("Song/Release is currently not in your library.");
+            return false;
+        }
     }
 
 }
